requirements.txt
Django>=4.2,<6
dj-database-url>=2.1
psycopg2-binary>=2.9
djangorestframework
joblib
numpy
Pillow
scikit-learn
